requirements.txt
# Core agent orchestration
langchain>=0.3.14
langchain-core>=0.3.14
langchain-community>=0.3.14
langchain-ollama>=0.2.0
langchain-text-splitters>=0.3.4
langchain-openai>=0.2.14
langgraph>=0.2.55

# Embeddings / vector store
faiss-cpu>=1.8.0

# LLM client (optional; primary path is Ollama via langchain-community)
openai>=1.59.0

# Config
python-dotenv>=1.0.1

# UI
streamlit>=1.33.0

# PDF text extraction (uploads)
pypdf>=5.0.0

# Structured plans / schemas
pydantic>=2.0.0

# PostgreSQL metadata
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
